Add --unsigned_payload_only flag to apexer
The flag instructs apexer to output the payload before it is signed by
avbtool.

The CL also contains new test to ensure that when the unsigned payload
is signed by avbtool, it is same as the payload we get when we unzip an
apex.

Bug: 149993331
Test: atest --host apexer_test
Change-Id: I0c0c29ee2fabaa447f5844e71949a00778b1d935
Merged-In: I0c0c29ee2fabaa447f5844e71949a00778b1d935
(cherry picked from commit 2cc12eb55f0f322397870f344c9806a959365373)