No longer disable com.android.support.apexer conditionally
This fixes the previous mistake that the APEX was disabled for
linux_bionic OS, which broke one of the ART builds. At the time it was
disabled with the hope that we don't build it for the new
linux_bionic_arm64 target. But we disabled it for all linux_bionic_*
targets. The ART build was using the linux_bionic_x86_64 target.

Enabling the linux_bionic_x86_64 target doesn't work because APEXes are
multilib:common modules. Enabling or disabling it for a specific
architecture is not possible.

In fact, the root cause of the problem was that the APEX build system
tried to include artifacts from all the host targets that are available.
That issue is fixed by skipping the host targets which can't be
executed natively on the host machine (i.e. HostCross == true).

Bug: 169454252
Test: HOST_CROSS_OS=linux_bionic HOST_CROSS_ARCH=arm64 m
Test: DIST_DIR=$(pwd)/out/dist \
        ./art/tools/dist_linux_bionic.sh com.android.support.apexer
Change-Id: If0a29fc2583883f2fb84e85fd25ae0200ee763e3