Adding multiple versions of a shared library.
This change adds a new shared library. The build_artifact.sh script
dynamically change the behaviour of the shared library so we can
confidently test that the apex under test is being linked to the
expected version of the library.

We are also preparing multiple versions and combination of test apex +
shared library in preparation for the extension of tests.

No new tests were added by this CL, they will be added separately.

Bug: 161542925
Test: atest sharedlibs_host_tests
Change-Id: I1c70c45db49fd8a0c2107842653d32d02506dbb8