Use better way for collecting staged session info from host side
When collecting staged session info using `pm get-stagedsessions`, we
can no longer assume that session info will fit in one line. Current
test assumes that if a line contains the session id, that line contains
all the information of that session. This is wrong.

We already have a method that can handle multiple line output for staged
session. Improved the logic a bit and updated all tests to use the
method.

On follow up CL, will move the logic to utility library.

Bug: 172562340
Test: atest ApexRollbacktests#testRebootingOnceDoesNotRevertSession
Test: atest ApexRollbackTests
Change-Id: I9eccc620dbad8b095ea88d70351281e0062932cf