Manage fs config for library directories
Test binary from sharedlibs test APEX was not able to be executed
with non-root environment. This was because symlinked library path did
not have execute permission. Permission was set from repack script, but
previous permission when the path was a file was not removed properly.
This change clears some previous configurations before adding new
configurations.

Bug: 172933725
Test: foo_test executed from non-root environment
Change-Id: I10cf9971b13beb23ee1961f843a80e084771cbb8