ApexFile::GetPath() returns realpath
When an APEX file is on a symlink directory (e.g. /system_ext/apex which
actually is /system/system_ext/apex on a device where system_ext is not
a real partition), the apex info constructed by scanning the directory
can be different from apex info re-constructed by reading /proc/mounts.
The latter shows realpath, while the former isn't.

This can make getAllPackages() incorrect. It is a union of the active
apexes gathered from /proc/mounts and factory apexes gathered by
scanning the built-in directories.

Fixing the bug by storing realpath of the apex file to the ApexFile
struct.

Bug: 179211712
Test: ApexTestCases.ApexServiceTest#GetAllPackages
Test: ApexTestCasea.ApexFileTest#GetPathReturnsRealpath
Change-Id: I5da96f78ddb11609c1271edfd1044d1979180b3c