Simplify clean up of APEX on boot complete
After the refactoring of APEX boot sequence, we no longer activate an
APEX if it doesn't have corresponding pre-installed version on device.
As such, there won't be any dangling APEX mounts.

We however can have APEX on data that are inactive. For example, after
an OTA if the pre-installed copy has higher version than the one on
data, then pre-installed copy will be activated. We don't clean up the
copy on data immediately, since that might corrupt non-fs-checkpointing
devices. So we wait till boot is completed and clean them up. Any data
apex which has not been activated gets deleted.

Bug: 172911820
Bug: 178750591
Test: atest ApexTestCases StagedInstallTest StagedInstallInternalTest
Test: atest sharedlibs_host_tests apexd_host_tests
Change-Id: I750b0196cf07ad6e8430b0d00da121aa48f2c404