Make apexd consider decompressed APEX as factory packages
Currently, we only consider APEX that are located in built-in
directories to be factory packages. This provides wrong result for
decompressed APEX which in located in /data parition.

In this CL, we first define a method to identify decompressed APEX. We
consider an APEX to be decompressed if it is a hardlink from
/data/apex/decompressed directory.

With this definition, we now consider an APEX to be factory package if
it is located in built-in directory or it is a decompressed APEX.

We had to update GetFactoryPackages a bit so that we don't include both
compressed and decompressed version of package X when searching
for factory packages.

We also had to update how we select APEX for activation. Previously,
when two APEX had same version, we gave the non-pre-installed copy
priority. But now, we can have two APEX with same version and both
pre-installed (compressed and decompressed). We select decompressed
version.

Bug: 172911820
Test: atest ApexCompressionTests#testDecompressedApexSurvivesReboot
Test: atest ApexCompressionTests#testDecompressionDoesNotHappenOnEveryReboot
Change-Id: I912ba0ca0bcf7b8f70298d4d8435e03614197fdd