Make calculation of size required for apex decomperssion more accurate
We are currently simply summing up all the sizes that update_engine is
providing us. However, apexd might not decompress every capex it finds
inside the OTA.

In this CL, we make apexd compare against the current apex installed
on device to make a more accurate calculation of which new apex
should be decompressed.

Bug: 172911822
Test: atest ApexdUnitTest ApexFileRepoTest ApexTestCases
Change-Id: I6f3c4c5aa4d800dd755c7953ec2d15327966874a