Set up environment for testing apex mount/unmount logic
This environment will make it much easier to write more tests in
apexd_test.cpp. As first use case, I'll use it for testing a command
that apexd is going to provide for otapreopt_chroot to setup apex
environment for otapreopt.

Since the test environment set up itself is tricky, I've also added test
for the test environment itself.

Bug: 181182967
Test: atest ApexTestCases
Change-Id: I4da8a43f282ba89ea234177245634b86bca2e7a7