Take into account apexes on /data for --otachroot-bootstrap
The policy is the same as for regular apexd flow: if update on /data has
same or higher version, it get's picked over the pre-installed version.

Implementation ended a little bit messy, since some of the apexd
internals are still not fully migrated to ApexFileRepository. I will
clean it up in a follow up cl.

Test: atest ApexTestCases
Bug: 181182967
Change-Id: Ic743c11f180c4ec74d84e57597a6b21dc40eb436