Fallback to pre-installed version if failed to activate /data one
To speed up boot flow, in S apexd was changed to first figure out what
apexes to activate, and then activate them in parallel. While it works
for the most cases, it unfortunately missed a case when an apex on /data
for some reason failed to activate. If that happened, device would end
up in a weird state when no version of apex is activated.

We address this issue by figuring out list of /data apexes that failed
to activate and falling back to activating corresponding pre-installed
apexes.

Since I was here, also simplified logic in ActivateApexePackages: there
is no need to check if there a higher version of an apex already
activated, since we are activating all the apexes in a single sweep.

In a follow up change I will implement the same functionality for
otachrot apexd.

Bug: 183030903
Test: atest apexd_host_tests
Test: atest ApexTestCases
Change-Id: I329b562e1d60108ef004da48ef700010fc7daa22