Handle CAPEX when falling back to pre-installed APEX activation
If data apex fails to activate for any reason, we fall back to
activating pre-installed APEX.

This CL handles the scenario where the pre-installed APEX is
compressed and needs to be decompressed first.

It also handles the case where we fallback to a capex which
already has a decompressed version on data. This happens when
device is running on decompressed APEX and we try to install an
update. On rare occasion where we fail to activate the update and
also fail to revert the update, we will hit this case.

There is another case which has not been handled yet: if corrupt
data APEX has same version has CAPEX, then we will currently fail
to decompress CAPEX due to name conflict. This will be fixed on
follow up CL.

Bug: 183920905
Test: atest ApexTestCases ApexCompressionTests
Change-Id: I203cbc3ae3915f8818235f91fdfc4375b264a14e