Avoid name collision with data APEX during decompression
When activating missing apexes, we might need to decompress a
CAPEX that has same version as the corrupt APEX. In the
current code, that will lead to a naming conflict.

We could resolve this by either overwriting the corrupt APEX,
or use a different naming convention for decompressed APEX.

I decided to use the less agressive option out of the two, since
our system doesn't require we uphold the naming convention strictly.
Decompressed APEX will now have ".decompressed.apex" suffix.

This have the added benefit of identifying decompressed apex by the
naked eye.

Bug: 183920905
Test: atest ApexTestCases ApexCompressionTests
Change-Id: I16cb70fd858621a27cd9e39253d6f2f40d67b9a1