Check for harlinked file's existance before skipping decompression
Originally, we were only checking if decompressed APEX already exited
in /data/apex/decompressed folder to decide if we should skip
decompression and just directly use the hardlinked version of the apex.

We assumed that when decompressed apex existed in decompression
directory, the hardlinked apex will also exist. This assumption broke
when as part of test cleanup we deleted everything in /data/apex/active,
but left decompression directory intact.

In this CL, we now check for existance of harlinked apex too before
skipping decompression.

Bug: 185993972
Test: atest ApexTestCases:ApexdUnitTest
Change-Id: I958ed57fce7ee155858c8ec643c734ca3775e97b