Add root digest of original_apex into manifest of CAPEX
When a CAPEX is updated on system, we need to discard the previous
decompressed APEX on /data and re-decompress. We usually check
version of decompressed APEX against the version of CAPEX in
/system to determine if CAPEX has been updated or not.

While this works in production, it fails in some development
scenarios, e.g, when developers push updated capex to system without
version bump or when we send updated capex via droidfood OTA without
version bump. Basically, any update to system CAPEX without version
bump makes it impossible to differentiate by checking versions only.

All APEX containers have a root digest attached to its footer. In
this CL, we propagate that information into the manifest of CAPEX.
We will later use this digest to detect if CAPEX has been updated
or not.

Bug: 185708645
Test: atest --host apex_compression_test
Test: atest ApexTestCases:ApexManifestTest
Change-Id: Ic6178d1fe5538769d509f6c2294baba601a6cdb0