Generate tinyxml2-based Apex Info sources
We can reduce the size of some binaries in ART by switching from
libxml2 to libtinyxml2.

This change adds the use of the new -t/--tinyxml flag in xsdc to generate the
Apex Info List sources.

Since these sources are used from several components and the change targets
only ART components, for now there will be 2 versions generated: one with
dependency on libtinyxml2, for ART, and the other, as until now, depending on
libxml2, for all the other components.

Bug: 177209057
Test: m
Change-Id: Ifbe7667cabb049cf0b79f3f5119a8319f9463992