Propagate the error message during apex activation to session object
When APEX activation fails, we currently just log it and while resuming
the session in StagingManager, if we notice that APEX session has failed
to activate, we simply ask to check logcat in the session error message.

Checking logcat is not always feasible. In this Cl, we now store the
error message that we log from apexd into the ApexSession proto. This
message is later sent to PackageManager via ApexSessionInfo aidl object.

Bug: 178192690
Test: atest ApexTestCases:ApexActivationFailureTests
Test: atest ApexTestCases

Change-Id: Ie5018e147fb1cb7160e605ad40431cf53adfc030