Deflake apexd_test#OnBootstrapCreatesEmptyDmDevices
It looks like kernel sometimes doesn't generate an uevent for
DM_DEV_CREATE ioctl, which in turns means that user space paths are not
created by uevented, which results in test failing.

While I am investigating this, chane the test to only assert that
devices are created in kernel and are in the SUSPENDED state.

Test: presubmit
Bug: 193462349
Change-Id: I9d527c02f4fe36c153b71c162006f433ce6a2ae6