Move compression-related tests from apexservice_test to apexd_test
apexservice_test require test apexes to be pre-installed on the device,
and only way to cleanup is brutally deleting directories under
/data/apex, which might impact other tests. So, instead we are moving
tests to apexd_test which should be more hermetic.

Note: we should probably move the code inside
reserveSpaceForCompressedApex binder function into apexd.h and add unit
tests for it as well.

Bug: 194998400
Test: atest ApexTestCases
Change-Id: Ie5bfa0742bb5189bdad5016c53d6287811680ca0