Move remaining activation-related tests to apexd_test
Note: for some of the tests (e.g. remountPackages ones), I've decided
not to move them to apexd_test just yet, since the future of
remoutPackages call is unclear.

apexservice_test require test apexes to be pre-installed on the device,
and only way to cleanup is brutally deleting directories under
/data/apex, which might impact other tests. So, instead we are moving
tests to apexd_test which should be more hermetic.

Test: atest ApexTestCases
Change-Id: I5218e8483bf482d835e42daed4c508de7e2c5ce5