Remove support for postInstallHook
This feature is not used, and most likely is broken (b/172376930).
Additionally, there are challenges around making sure the hooks will be
stable across multiple release.

If we have a valid usecases for postInstallHook-like functionality, then
we can figure out a way to do them (or just revert this change).

Bug: 172376930
Bug: 172606645
Test: atest ApexTestCases
Change-Id: I228decdd8efd3574f7620ea59837be309ab96621