Wait for reboot or boot to complete in WaitForBootStatus call
We do not expect apexd to execute commands that follow a Reboot call.
But on occassions when reboot takes too long, we can end up executing
follow up statements and putting device in bad state.

This CL forcefully waits for an arbitrary amount of seconds for reboot
to complete and upon time out crashes the apexd to avoid executing
further statements.

Similarly, we do not expect apexd to break from WaitForBootStatus before
boot is complete. On occassions when there is a native crash and the
revert_fn method fails for any reason, we used to break out prematurely.
We now wait till boot is complete before breaking out.

Bug: 196203300
Test: atest ApexRollbackTests#testReasonForRevertIsLoggedDuringBootloop
Change-Id: If2aa272b8a2bb08d80968d68f6a52190544c4c74