Determine if staged apex contributes to classpath or not
In order to determine if staged apex contributes to classpath or not, we
need to mount them first so that we can look inside them. Once mounted,
defer the calculate of classpath to derive_classpath binary.

A new clas has been added to handle logics related to classpath. Unit
test for this class will be added on follow up CL.

Bug: 187444679
Test: atest ApexTestCases
Test: atest StagedInstallInternalTest
Change-Id: I3a6b39cca67c8971fd71c66d20202ec2b57b4c97