Tests for Apk-in-apex scanning
This test verifies that each apk is correctly associated with the
partition in which the containing apex is located.

This affects a number of features including the where PermissionManager
needs to look for the allowlist of privileged permission.

This test was written to exhibit the bug that got fixed by
I690618727ebf9a61f07b3925b30e7631d33c4e8e

Test: atest
Change-Id: I4cec3fdf05f2ab1a3b5032870a74562aed53cb95