Temporarily disable apexer_test during JDK udpate
The lab machines that run host unit tests have a snapshot of
the platform JDK in their docker image.  When the platform
JDK is updated there is a lag until the lab docker image picks
up the update and is deployed.  apexer_test runs signapk on
the lab machines, and is getting different results when the
JDK doesn't match.  Temporarily disable automatically running
the test until the lab JDK is updated.

Bug: 205944740
Bug: 210142956
Test: treehugger
Change-Id: I1a0ccab5b1b8fbc5fb2603c289701cf268d2cc44