Specify rotation target for CTS shims
Android T introduced the v3.1 signature scheme to allow targeting of
APK signing key rotation. This new scheme is defaulting signing key
rotation to target T; for those packages that previously rotated,
the rotation-min-sdk-version parameter can be specified to maintain
the previous behavior of targeting P for rotation. This commit
updates the apksigner commands for the CTS shims with this new
parameter to maintain the previous behavior and not require the
original signing key for the V3 signing block.

Bug: 230486681
Test: mmm system/apex/shim/build/
Ignore-AOSP-First: New apksigner API is only available in internal.
Change-Id: I8c21c1f1682d6a5de9089634afed3e12b02dd4df