Add CtsShim variants for testing <install-constraints>.
CtsShim.apex is pre-install on all devices and they can be used to test
PackageManager behaviour in CTS, as long as the test apexes do not
contain code.

The artifacts cannot be immediately used yet, because the prebuilts
of these variants need to be signed and dropped into gerrit before
CTS can use these artifacts.

Bug: b/232475788
Test: presubmit + manual local tests
Ignore-AOSP-First: tm-qpr-dev change first, before cp-ing to AOSP
Change-Id: I4ec60221512289c45103c22b611dfe76c4da66c6